8763: show unserved petition warning
and remove 'serve' button when necessary
